Use DonateCoordinator for article campaign donate flow
